oi vey - more server vs client side cleanup
